requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
PyPDF2
openpyxl
anthropic
python-dotenv
pydantic
apscheduler
